refactor(contracts): one-type-per-file resolver; seal test classes; docs lint
Move WorkflowCatalogResolution into its own file (one public type per
file; keeps INV-6 sealed / INV-7 init-only). Seal the new SMQ test classes
(WorkflowCatalogTests + IntentEnvelope/WorkflowRef/MergeGateDecision/
JourneyResult tests sealed in their respective commits) per repo
sealed-by-default. Backtick the #99/#101/#102/#103 issue-ref so it does not
render as a heading; tag the PR-flow fenced block as text.
