+++ docs/designs/2026-05-24-v2.8.0-closeout.md
@@ -10,7 +10,7 @@
 ## 1. Problem
 
 Slice A (`Strategos.Contracts` 0.2.0 — events / workflow IR / invariant catalog) shipped via
-#99/#101/#102/#103. The 2.8.0 milestone has 10 open issues that, taken together, are the
+`#99/#101/#102/#103`. The 2.8.0 milestone has 10 open issues that, taken together, are the
 *remaining* close-out surface. They split into three coherent units plus two administrative
 dispositions — a deliberate **3-PR close-out**, not ten independent tasks.
 
@@ -126,7 +126,7 @@ Three small follow-ups closing drift surfaces left by #51/#52, plus the #54 clos
 
 ## 7. Sequencing & invariant summary
 
-```
+```text
 PR 1 (S2→S3→S4→S1)  ──┐
 PR 2 (#105,#106,#107) ─┼──►  milestone close  (defer #100, close #54)
 PR 3 (#104 bridge)  ──┘
